Moved dialogs to new folder
