+++ src/renderer/pages/ConfigPage.tsx
@@ -5,13 +5,13 @@ import { ClusterType } from 'models/Cluster'
 import { useState } from 'react'
 import { ReflexContainer, ReflexElement, ReflexSplitter } from 'react-reflex'
 import 'react-reflex/styles.css'
-import AlertDialog from 'renderer/common/AlertDialog'
 import PageRoot from 'renderer/common/PageRoot'
-import ConfigurationDialog from 'renderer/components/ConfigurationDialog'
 import GitView from 'renderer/components/GitView'
 import LogsView from 'renderer/components/LogsView'
-import SettingsDialog from 'renderer/components/SettingsDialog'
 import StatusView from 'renderer/components/StatusView'
+import AlertDialog from 'renderer/dialogs/AlertDialog'
+import ConfigurationDialog from 'renderer/dialogs/ConfigurationDialog'
+import SettingsDialog from 'renderer/dialogs/SettingsDialog'
 import { ConfigFileService, useConfigFileState } from 'renderer/services/ConfigFileService'
 import { DeploymentService, useDeploymentState } from 'renderer/services/DeploymentService'
